Send a pull request for Github
If you are using Megam and a contributor, this would help you to send us a pull request to Megam.
First step is to setup the master pointing to your_userid_repo and origin to megamsys/
Here we demonstrate by changing a README.md file and push the change to master.

Upon commit, open up github.com//. Read this to send a pull request.
https://help.github.com/articles/using-pull-requests/ ( https://help.github.com/articles/using-pull-requests/ )